Use common download function and predownload appimage tools (#708)
* Move appimage tool downloading to install_packages_root like gitversion

* Add missing package for Ubuntu

* Fix GITHUB_TOKEN

+++ ShellScripts/Linux/install_gitversion_fn.sh
@@ -8,16 +8,18 @@ install_gitversion() {
     fi
 
     local script_dir=$(cd "$(dirname "${BASH_SOURCE[0]}")" &> /dev/null && pwd)
-    pushd "$script_dir"
-
-    source ../common/download_github_release_fn.sh
+    source "$script_dir/../common/download_github_release_fn.sh"
 
     local gitversion_tgz
     download_github_release gitversion_tgz "GitTools" "GitVersion" "linux-x64" "$outputdir"
-    tar xfz ${gitversion_tgz} --directory "$outputdir"
+    if ! tar xfz ${gitversion_tgz} --directory "$outputdir"; then
+        echo "❌ Could not unpack gitversion tgz!" >&2
+        return 1
+    fi
     chmod +x "$outputdir/gitversion"
-    mv "$outputdir/gitversion" /usr/local/bin/gitversion
+    if ! mv "$outputdir/gitversion" /usr/local/bin/gitversion; then
+        echo "❌ Could not move gitversion to /usr/local/bin!" >&2
+        return 1
+    fi
     rm -f ${gitversion_tgz}
-
-    popd
 }
